requirements.txt
# Core dependencies
streamlit>=1.28.0

# LlamaIndex core and components
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-readers-file>=0.1.0

# Transformers and model dependencies
transformers>=4.30.0
torch>=2.0.0

# Additional utilities
python-dotenv>=1.0.0
